[TWINFACES-189] added Experimental_CrudDataTable for table statuses
