+++ src/entities/twinStatus/libs/index.ts
@@ -1,2 +1,3 @@
 export * from "./hooks";
 export * from "./constants";
+export * from "./types";
