+++ src/entities/twinStatus/libs/types.ts
@@ -1,0 +1,6 @@
+import { z } from "zod";
+import { TWIN_CLASS_STATUS_SCHEMA } from "@/entities/twinStatus";
+
+export type TwinClassStatusFormValues = z.infer<
+  typeof TWIN_CLASS_STATUS_SCHEMA
+>;
